# Reduce memcpy serialization