--- a/protocol/src/lib.rs
+++ b/protocol/src/lib.rs
@@ -6,6 +6,7 @@ use anyhow::bail;
 use bytes::{Buf, BufMut, BytesMut};
 use num_enum::IntoPrimitive;
 use serde::{Deserialize, Serialize};
+use strum_macros::EnumDiscriminants;
 use tokio_util::codec::{Decoder, Encoder};
 use uuid::Uuid;
 
@@ -282,7 +283,10 @@ pub const CRUCIBLE_MESSAGE_VERSION: u32 = 5;
  * go do that right now before you forget.
  */
 #[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[derive(
+    Debug, PartialEq, Clone, Serialize, Deserialize, EnumDiscriminants,
+)]
+#[strum_discriminants(derive(Serialize))]
 #[repr(u16)]
 pub enum Message {
     /**
@@ -531,6 +535,8 @@ pub enum Message {
     /*
      * IO related
      */
+    // Message::Write must contain the same fields in the same order as
+    // RawMessage::Write which is used for zero-copy serialization.
     Write {
         upstairs_id: Uuid,
         session_id: Uuid,
@@ -580,6 +586,8 @@ pub enum Message {
         responses: Result<Vec<ReadResponse>, CrucibleError>,
     },
 
+    // Message::WriteUnwritten must contain the same fields in the same order as
+    // RawMessage::WriteUnwritten, which is used for zero-copy serialization.
     WriteUnwritten {
         upstairs_id: Uuid,
         session_id: Uuid,