--- a/upstairs/Cargo.toml
+++ b/upstairs/Cargo.toml
@@ -17,6 +17,7 @@ anyhow.workspace = true
 async-trait.workspace = true
 async-recursion.workspace = true
 base64.workspace = true
+bincode.workspace = true
 bytes.workspace = true
 chrono.workspace = true
 crucible-common.workspace = true