Right now, we copy data from a `BlockOp::Write/WriteUnwritten` into a `BytesMut` buffer, encrypt it in-place, then serialize it using `CrucibleEncoder`.  Serialization requires a *second* full copy of the data.  This is noticeable in our flamegraphs for large writes:

![before](https://github.com/oxidecomputer/crucible/assets/745333/901a8a68-578b-4824-8d37-ab0ccc73ec6a)
[interactive version](https://github.com/oxidecomputer/crucible/assets/745333/901a8a68-578b-4824-8d37-ab0ccc73ec6a)

This PR performs encryption and serialization simultaneously into a single buffer, saving us the unnecessary copy:

![after](https://github.com/oxidecomputer/crucible/assets/745333/e8720f2d-1ed8-4c86-b706-671fe28566de)
[interactive version](https://github.com/oxidecomputer/crucible/assets/745333/e8720f2d-1ed8-4c86-b706-671fe28566de)

This is a significant speedup for large writes.  Comparing against `main` (best of 2x runs), I see the following:

| Random write size | `main` | `reduce-memcpy-serialization` |
|-|-|-|
| 4K | 27.0 MiB/s | 27.1 MiB/s |
| 1M | 710 MiB/s | 861 MiB/s |
| 4M | 890 MiB/s | 969 MiB/s |

The implementation introduces new `IOop::SerializedWrite` and `IOop::SerializedWriteUnwritten` variants, which represent pre-serialized data.  Unlike normal `IOop`s (which are translated into the existing `enum Message`), these are translated into a new `RawMessage` type.  The serialization must be equivalent to the existing `Message::Write/WriteUnwritten` values; this is checked implicitly in our unit tests, which decode into the `Message` type.