@mkeeter 2024-01-08
> Curious, why keep the old IOop::Write and make a new IOop::SerializedWrite? Don't we always want to do the serialization for Wrrite/WU?
> 
> Is it because what happens in the downstairs still needs things in IOop::Write format? Ah, no, I think I see now. Because we want to encode write/WU outside

Yup, the issue here is that the same `IOop` type is used by both `crucible-upstairs` and `crucible-downstairs`.

Thinking about it a little more, I'd vote for splitting it into separate types (i.e. copy the current state of `enum IOop` into `crucible-downstairs`), and make the type private to each crate.  That would let us make (upstairs) `IOop::Write/WriteUnwritten` pre-serialized (which is what you're suggesting), and we could later make (downstairs) `IOop::ReadResponse` pre-serialized in the same way.