@mkeeter 2024-01-08
@leftwo a1a535efad1c2b6046d22a555a6488f5d846ee3a splits the `IOop` into Upstairs vs Downstairs flavors, and also has the benefit of removing impossible variants from each one.