@mkeeter 2024-01-08
d34d4b76c57fb97d4d508108edc82564d702ebcb adds unit tests to make sure that a `RawMessage` deserializes to the correct `Message`.  It's a _little_ manual, so I'm open to suggestions for ways to clean it up (but I don't think it's bad enough to block).